refactor(node-red-s1seven-api): restructure nodes based on API resources
