+++ packages/node-red-s1seven-api/lib/certificates/locales/en-US/hash-certificate.json
@@ -1,0 +1,16 @@
+{
+  "certificates": {
+    "label": {
+      "apiConfig": " API config",
+      "name": " Name",
+      "algorithm": " Algorithm",
+      "encoding": " Encoding"
+    },
+    "errors": {
+      "accessToken": "Please add an access token",
+      "algorithm": "Please choose a valid algorithm option",
+      "encoding": "Please choose a valid encoding option",
+      "validCertificate": "Please add a valid JSON certificate to msg.payload or global.certificate"
+    }
+  }
+}
