+++ packages/node-red-s1seven-api/lib/certificates/locales/en-US/verify-certificate.json
@@ -1,0 +1,20 @@
+{
+  "certificates": {
+    "label": {
+      "apiConfig": " API config",
+      "name": " Name",
+      "algorithm": " Algorithm",
+      "encoding": " Encoding",
+      "identity": " Identity",
+      "mode": " Mode"
+    },
+    "errors": {
+      "accessToken": "Please add an access token",
+      "algorithm": "Please choose a valid algorithm option",
+      "encoding": "Please choose a valid encoding option",
+      "validCertificate": "Please add a valid JSON certificate to msg.payload or global.certificate",
+      "companyId": "Please add a company id",
+      "identity": "Please add an identity"
+    }
+  }
+}
