+++ packages/node-red-s1seven-api/lib/companies/locales/en-US/get-company.html
@@ -1,0 +1,88 @@
+<script type="text/html" data-help-name="get company">
+  <p>A simple node that returns a company object from S1Seven's servers</p>
+
+  <h3>Inputs</h3>
+  <dl class="message-properties">
+    <dt class="message-properties">
+      accessToken <span class="property-type">string</span>
+    </dt>
+    <dd>
+      the access token to use to authenticate the request. This must be
+      generated using the <code>get access token</code> node.
+    </dd>
+    <dt class="message-properties">
+      companyId <span class="property-type">string</span>
+    </dt>
+    <dd>the id of the company to associate with the notarization.</dd>
+    <dt class="optional">
+      environment <span class="property-type">staging | production</span>
+    </dt>
+    <dd>
+      choose the environment that you want to use - app.s1seven.dev or
+      app.s1seven.com. Defaults to production (app.s1seven.com).
+    </dd>
+    <dt class="optional">version <span class="property-type">1</span></dt>
+    <dd>
+      choose the version of the api to use. It can be set in the config node.
+      Defaults to <code>1</code>
+    </dd>
+  </dl>
+
+  <h3>Outputs</h3>
+  <ol class="node-ports">
+    <li>
+      Success output
+      <dl class="message-properties">
+        <dt>payload <span class="property-type">object</span></dt>
+        <dd>the response from the HTTP request containing a company object.</dd>
+        <dt>headers <span class="property-type">object</span></dt>
+        <dd>the headers from the HTTP response.</dd>
+      </dl>
+    </li>
+    <li>
+      Failure output
+      <dl class="message-properties">
+        <dt>payload <span class="property-type">object</span></dt>
+        <dd>the error response from the HTTP request.</dd>
+        <dt>headers <span class="property-type">object</span></dt>
+        <dd>the headers from the HTTP response.</dd>
+      </dl>
+    </li>
+  </ol>
+
+  <h3>Details</h3>
+  <p>
+    The following properties can be added via the admin ui, found by
+    double-clicking on the hash node, or by passing them in using the
+    <code>msg</code> object. If they are passed in via the
+    <code>msg</code> object, they will have priority over the values found in
+    the admin ui. This allows the values to be overridden dynamically if
+    necessary. For example, if both <code>msg.companyId</code> and a companyId
+    added via the admin ui are present, <code>msg.companyId</code> will be used
+    instead of the admin ui access token.
+  </p>
+
+  <p>
+    <code>accessToken</code> is used to authenticate the HTTP request. To find
+    out how to generate an access token, look at the documentation at
+    <a>https://manual.s1seven.com</a> under the developers section. It can be
+    input via <code>msg.accessToken</code>, or via
+    <code>global.s1sevenAccessToken</code>, in that order of priority.
+  </p>
+
+  <p>
+    <code>companyId</code> is the id of the company to be associated with the
+    notarization. To find out how to get a company id, look at the documentation
+    at <a>https://manual.s1seven.com</a> under the developers section. It can be
+    input via <code>msg.companyId</code>, via the ui which can be found by
+    double-clicking on the <code>get company</code> node, or via
+    <code>global.companyId</code>, in that order of priority.
+  </p>
+
+  <p>
+    <code>msg.environment</code> allows you to pass in the string value
+    <code>staging</code> or <code>production</code> to choose which server you
+    wish to use. Learn more about our environments here:
+    <a>https://developers.s1seven.com/docs/environments/</a>.
+  </p>
+</script>
