--- packages/node-red-s1seven-api/lib/company/locales/en-US/company.json
@@ -1,14 +1,0 @@
-{
-  "company": {
-    "label": {
-      "apiConfig": " api config",
-      "name": " Name"
-    },
-    "errors": {
-      "accessToken": "Please add an access token",
-      "companyId": "Please add a company id",
-      "identity": "Please add an identity",
-      "validCertificate": "Please add a valid JSON certificate to msg.payload or global.certificate"
-    }
-  }
-}
