--- packages/node-red-s1seven-api/lib/identities/locales/en-US/identities.json
@@ -1,18 +1,0 @@
-{
-  "identities": {
-    "label": {
-      "apiConfig": " api config",
-      "name": " Name",
-      "status": " Status",
-      "coinType": " BIP44 coin type",
-      "BIP44Account": " BIP44 Account number",
-      "BIP44Index": " BIP44 Index number"
-    },
-    "errors": {
-      "accessToken": "Please add an access token",
-      "companyId": "Please add a company id",
-      "identity": "Please add an identity",
-      "validCertificate": "Please add a valid JSON certificate to msg.payload or global.certificate"
-    }
-  }
-}
