--- packages/node-red-s1seven-api/lib/notarize/locales/en-US/notarize.json
@@ -1,15 +1,0 @@
-{
-  "notarize": {
-    "label": {
-      "apiConfig": " api config",
-      "name": " Name",
-      "identity": " Identity"
-    },
-    "errors": {
-      "accessToken": "Please add an access token",
-      "companyId": "Please add a company id",
-      "identity": "Please add an identity",
-      "validCertificate": "Please add a valid JSON certificate to msg.payload or global.certificate"
-    }
-  }
-}
