--- packages/node-red-s1seven-api/lib/notarize/notarize.html
@@ -1,44 +1,0 @@
-<script
-  type="text/javascript"
-  src="/resources/@s1seven/node-red-s1seven-api/constants.js"
-></script>
-<script type="text/javascript">
-  RED.nodes.registerType('notarize certificate', {
-    category: 'S1Seven',
-    color: '#BAD3FB',
-    defaults: {
-      name: { value: '' },
-      apiConfig: { value: '', type: 'api-config' },
-      identity: { value: '', required: false },
-    },
-    inputs: 1,
-    outputs: 2,
-    outputLabels: OUTPUT_LABELS,
-    icon: 's_icon_logo.svg',
-    label: function () {
-      return this.name || 'notarize certificate';
-    },
-  });
-</script>
-
-<script type="text/html" data-template-name="notarize certificate">
-  <div class="form-row">
-    <label for="node-input-name"
-      ><i class="fa fa-tag"></i><span data-i18n="notarize.label.name"></span
-    ></label>
-    <input type="text" id="node-input-name" placeholder="Name" />
-  </div>
-  <div class="form-row">
-    <label for="node-input-apiConfig"
-      ><i class="fa fa-tag"></i
-      ><span data-i18n="notarize.label.apiConfig"></span
-    ></label>
-    <input type="text" id="node-input-apiConfig" placeholder="api-config" />
-  </div>
-  <div class="form-row">
-    <label for="node-input-identity"
-      ><i class="fa fa-tag"></i><span data-i18n="notarize.label.identity"></span
-    ></label>
-    <input type="text" id="node-input-identity" placeholder="identity" />
-  </div>
-</script>
