+++ packages/node-red-s1seven-api/lib/tokens/locales/en-US/get-token.json
@@ -1,0 +1,12 @@
+{
+  "tokens": {
+    "label": {
+      "apiConfig": " API config",
+      "name": " Name"
+    },
+    "errors": {
+      "clientId": "Please add a client id",
+      "clientSecret": "Please add a client secret"
+    }
+  }
+}
