--- packages/node-red-s1seven-api/lib/tokens/locales/en-US/tokens.json
@@ -1,12 +1,0 @@
-{
-  "tokens": {
-    "label": {
-      "apiConfig": " api config",
-      "name": " Name"
-    },
-    "errors": {
-      "clientId": "Please add a client id",
-      "clientSecret": "Please add a client secret"
-    }
-  }
-}
