--- packages/node-red-s1seven-api/lib/verify/locales/en-US/verify.json
@@ -1,12 +1,0 @@
-{
-  "verify": {
-    "label": {
-      "apiConfig": " api config",
-      "name": " Name",
-      "mode": " Mode"
-    },
-    "errors": {
-      "validCertificate": "Please add a valid JSON certificate to msg.payload or global.certificate"
-    }
-  }
-}
